Move lots of tests around, and tidy up the pom.
