+++ core/src/test/java/org/truth/gwt/TruthTest.gwt.xml
