+++ core/src/test/java/org/truth/subjects/IntegerTest.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.truth;
+package org.truth.subjects;
 
 import static org.junit.Assert.fail;
 import static org.truth.Truth.ASSERT;
@@ -25,6 +25,7 @@ import org.junit.Test;
 import org.junit.internal.AssumptionViolatedException;
 import org.junit.runner.RunWith;
 import org.junit.runners.JUnit4;
+import org.truth.Expect;
 
 /**
  * Tests for Integer Subjects.
