+++ core/src/test/java/org/truth/subjects/StringTest.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.truth;
+package org.truth.subjects;
 
 import static org.junit.Assert.fail;
 import static org.truth.Truth.ASSERT;
